requirements.txt
matplotlib==3.5.1
numpy~=1.23.4
pandas==1.4.3
Pillow==9.3.0
seaborn==0.11.2
tqdm==4.64.0

scikit-image~=0.19.3
keras~=2.10.0
scikit-learn~=1.1.3
tensorflow~=2.10.0